requirements.txt
awscli==1.7.24
beautifulsoup4==4.3.2
boto==2.36.0
botocore==0.105.0
bson==0.4.6
Django==1.10
django-crispy-forms==1.3.2
django-extra-views==0.6.3
django-haystack==2.5.0
django-imagekit==3.0.4
djangorestframework==3.5.4
django-webpack-loader
Fabric==1.13.1
gunicorn==0.17.4
elasticsearch==5.1.0
certifi==2016.9.26
requests
six==1.10.0
Unidecode==0.4.19
uWSGI==2.0.12
virtualenv==12.1.1
django-polymorphic==1.0.2
psycopg2
